Rename compute_tasks folder to computeTasks
